[ENH] `_safe_import` to allow multiple inheritance from multiple mock classes (#8061)
In https://github.com/sktime/sktime/issues/8055 and
https://github.com/sktime/sktime/pull/7869, @PranavBhatP reported that
`_safe_import` will fail under specific circumstances to behave like the
imported object if it is not available, namely:

* two classes are safe-imported, and
* another class is defined that inherits from both (directly or
indirectly)

This would fail, as the resulting class has `MagicMock` in its
inheritance tree twice.

This PR fixes #8055 by ensuring that classes imported by `_safe_import`
are dynamic, with the same behaviour as `MagicMock`, rather than
`MagicMock` itself.

A test is added with the minimal failure case of #8055.