[ENH] refactor `datatypes` registry to `scikit-base` records (#7816)
This PR refactors the `datatypes` registry to the new, programmatic
`scikit-base` records, deduplicating the points of extension.

The manually popultated records in the `_registry` folders are removed
in favour of the records currently in the `_check` folders.

In a future PR, we may rename `_check` to `_registry`, as these contain
the primary records now.