[ENH] automatic transformer coercion in explicit transformer pipeline specification syntax (#8160)
This PR adds automatic transformer coercion in transformer pipelines,
and removes the need for manual transformer coercions such as of
`sklearn` estimators through `TabularToTransformAdaptor`.

This is achieved through use of the `registry._scitype_coercion` module,
which also has minor updates and bugfixes in this PR - the coercion
logic is moved from the dunder to the construction of the stpes.

Complements https://github.com/sktime/sktime/pull/8153 for
transformation pipelines.