[ENH] Add `is_hierarchical` method to check whether a dataframe is hierarchical (#8170)
#### Reference Issues/PRs
Reference to [ENH] [#8112](https://github.com/sktime/sktime/issues/8112)


#### What does this implement/fix? Explain your changes.
Adds the `utils.multiindex.is_hierarchical` method that evaluates if a
dataframe's index is truly hierarchical. This is important because
certain algorithms assume that a dataframe is hierarchical, and you can
get unexpected behavior if it is not. For example,
`Reconciler(method='td_fcst')` will return `NaN` values if the dataframe
index is not truly hierarchical.