[ENH] enforce `MeanSquaredErrorPercentage` to be positive (#8318)
This changes `MeanSquaredErrorPercentage` to be positive if the
denominator - the mean of ground truth samples - is negative.

This also solves an unreported sporadic bug in rarer cases when, on the
test data, said denominator is negative, and the square root is computed
via `square_root=True`.

I do not consider this a breaking change, since sound usage should
happen only if the average is indeed a positive number - and since the
"lower is better" tag was set to `True`.