[ENH] Visualisation utility for global splits (#8204)
DEPENDS on #8064
#### Reference Issues/PRs

This plot visualises the different splits created by the folds used for
evaluating global forecasters.


#### What does this implement/fix? Explain your changes.

It extracts the `gen_y_X_train_test_global` so that it can be used in
other `sktime` modules. This returns the different train/test splits
used in evaluation. These splits are visualised in the newly added
function `plot_folds_global_forecasting` in `plotting.py`

#### Exemplary output:


![image](https://github.com/user-attachments/assets/3cd2639d-6336-459f-adb3-004465d0c58d)