[DOC] Add reference to the updated roadmap page in `contributing.rst` (#8350)
The [roadmap](https://www.sktime.net/en/latest/roadmap_2021.html) page
linked on the
[contributing](https://www.sktime.net/en/latest/get_involved/contributing.html)
page, leads to an outdated roadmap page which is not expected to be
there.

What I think the issue might be:

The new roadmap page is contained in `roadmap.rst` and the old one is
contained in `roadmap_2021.rst`, `contributing.rst` still references to
the old roadmap page. This PR attempts to fix it.