[ENH] Implement reconcilers as transformations, and add new non-negative reconciliation (#7697)
This PR adds new reconciliation methods and revisits the current
reconciliation API.

Currently, the reconciliation is done at `transform`. However, to be
able to have all reconciliation strategies under the same
API, we should notice that there are two important steps:

1. `transform`: Pre-process the original timeseries, by converting them
to shares (in Topdown-share), by keeping just the relevant ones
(bottom-levels, in bottom-up approach), or by doing nothing (optimal
reconciliation).
2. `inverse_transform`: Reconcile the base forecasts

In this PR, the zero-constrained representation was used to compute
optimal reconciliation because it is more efficient and more versatile
than the projection one.

#### Reference Issues/PRs

Add some of the methods mentioned in #7609 
Fixes #7614
Fixes #8393

#### What does this implement/fix? Explain your changes.

- Add Topdown (share) reconciliation
- Add optimal non-negative reconciliation (using cvxpy to solve
quadratic programming problem)
- Add middle-out reconciliation
- Add forecast proportions and optimal reconciliation (called
"FullHierarchyReconciliation") as transformations
- Add new tag `capability:hierarchical_reconciliation` so that tailored
unit tests can be executed

#### Does your contribution introduce a new dependency? If yes, which
one?

Yes, optinal `cvxpy` to solve quadratic programming problem in
non-negative reconciliation


#### What should a reviewer concentrate their feedback on?

- The new API design proposal
- The directory name, and we should discuss on how to handle deprecation
and fix tutorials
- The new tag

#### Did you add any tests for the change?

Yes, added tests for some utilities and for hierarchical reconciliation

#### Any other comments?

With these transformations implemented, it is easy to create or extend
the current ReconcileForecaster to account for these new possibilities.