[MNT] remove `stefanzweifel/git-auto-commit-action` from all-contributors workflow (#8369)
Removes the workflow dependency on the 3rd party GitHub workflow action
`stefanzweifel/git-auto-commit-action`, by replacing it with a variation
on a `git diff` call.

This is to enhance our security posture, as third party actions are
possible security vulnerabilities - even if pinned, they can be
compromised (unlike python packages), see for instance the `tj-actions`
compromise.