[ENH] test framework: "core classes" and non-core classes with different test trigger conditions (#8408)
To limit test load from the conditional tests, this PR introduces two
types of objects:

* "core objects", these trigger tests as currently, e.g., if the
framework modules are changed, these classes are also tested.
* "non-core objects", tests for these are only triggered if their code
directly changes, or a dependency changes.

The idea of "core objects" is to provide good coverage for integration
with framework concerns, and `sktime` native composition, while avoiding
to always trigger tests for all classes when framework contents change.

The designation as "core objects" happens through a boolean tag
`tests:core`.

By default, estimators are non-core (to make extension easy), and a
number of estimators are labelled manually as core, in this PR:

* all compositors for pipeline building or tuning
* all splitters, all metrics
* all dummy and naive estimators
* a selection of other estimators that covers a good (subjectively
choesn) range of estimator types and software concerns

All tests for all core estimators run in this PR, due to the change in
base class, which triggers a framework type run - so, test times in this
PR indicate how long full runs of all "core objects" take in the future.

Overall intended benefits:

* lower test run time
* avoiding test failures arising from a few hard-to-fix sporadic errors
* that, while not reducing de-facto test coverage by much

The framework changes themselves are tested via:
* https://github.com/sktime/sktime/pull/8411 to show that changed
estimators that are not core are still tested
* CI in this PR to check that core estimators get tested.