[ENH] Use `sktime` Provided Dataset in `TSC` Dataset Loaders (#8416)
There are two utility functions for loading TSC datasets in `sktime`,
`_load_dataset` and `_load_provided_dataset`. The former is for the
datasets that need to be downloaded and latter is for the datasets that
are shipped with sktime.

For example, `load_arrow_head` uses `_load_provided_dataset` but loaders
like `load_gun_point`, `load_osuleaf` and others use `_load_dataset`
even though they are shipped with sktime by default. Thus, it leads to
redundant dataset downloads. This PR fixes it.