[ENH] Add `ClassificationBenchmark` Class and its Helper Functions (#8357)
_Stacks on #8347 and #8347 stacks on #8267, please merge in that order._

This PR adds `ClassificationBenchmark` class for benchmarking of
classifiers. The design for it is mostly taken from the existing
`ForecastingBenchmark` class.

The benchmark run can be performed across a combination of estimators
and tasks, tasks can further be a combination of splitting strategies,
dataset, and scorers. The result of the benchmark run gets stored in a
file as well as returned as a `DataFrame`.

Below is an attached screenshot of an example run, the output has the
similar structure as that of the output from `ForecastingBenchmark`.

<img width="726" alt="Screenshot 2025-06-09 at 7 32 54 PM"
src="https://github.com/user-attachments/assets/95227da3-a652-4e02-b159-6215da7944ff"
/>