[ENH] enable univariate mode for `DynamicFactor` forecaster (#8345)
This PR adds a univariate mode for the `DynamicFactor` forecaster. While
it is degenerate in the sense of being full rank, this results in a
non-trivial smoothing model determined by the parameters.

The "trick" is to duplicate a univariate dataframe to one that is
bivariate. This will result in singular matrices but the algorithm still
produces valid output.

This is somewhat questionable in the modeling sense, but it simplifies
the interface by removing one of a very few number of forecasters that
do not have a univariate mode.

Reviewers should check whether this makes sense or whether this feels
too problematic.