[ENH] change default `capability:categorical_in_X` to `True` to enable composition case (#8231)
A few minor releases ago, the tag `capability:categorical_in_X` was
introduced to mark estimators that allow categorical features in `X`.

However, the tag was not appropriately set on compositors, so most
compositors do not let categorical data through - but they should.

Since checking and extending would be very labour intensive, I am
proposing to just set the default to `True`, and erring on the side of
letting data through that should not be let through, rather than
blocking data that should be let through.