[MNT] Delete Redundant Dataset Files (#8441)
TSC data loaders in `sktime` uses datasets in `.ts` format, but some of
the datasets that come pre-loaded with `sktime` contains same datasets
in 2-3 format. This is likely due to their original host
[https://www.timeseriesclassification.com](https://www.timeseriesclassification.com)
which contains `zip` of datasets in different formats.

This PR removes the redundant dataset files and reduces the repository
size by around 7.5 MB which is significant.