[BUG] fix forecasting metric functions incompatible with scikit-learn 1.7 #8457 (#8459)
#### Reference Issues/PRs
Fixes #8457. 

#### What does this implement/fix? Explain your changes.
Copies the internal v1.6 [_check_reg_targets from
sklearn](https://github.com/scikit-learn/scikit-learn/blob/main/sklearn/metrics/_regression.py#L61)
into sktime.utils.sklearn._metrics. This is a hotfix for forecasting
functions not being compatible with sklearn v1.7 which changed
_check_reg_targets' signature in this
[MR](https://github.com/scikit-learn/scikit-learn/pull/30886.). See
discussion on issue on the longer term fix.

This short term fix is probably worth it since _check_reg_targets is
called at the start of most forecasting._functions and there is no
workaround or kwarg that you can give to workaround the issue. It is a
hard TypeError in the signature.

#### What should a reviewer concentrate their feedback on?

This is a hotfix. I do not love the approach but I do not have time for
a more principled approach.


#### Did you add any tests for the change?

The existing tests should now catch this since the module
performance_metrics has changed. I am not happy that this was not
detected in the unit tests which is the actual cause of the bug - we
skip these performance metrics entirely if the module hasn't changed
from our side using our pytest `run_test_module_changed` mark. This
should ideally also rerun if dependencies changed but is outside the
scope of this MR.