[ENH] do not trigger VM tests for non-core estimators on framework changes (#8497)
PR https://github.com/sktime/sktime/pull/7087 works, but vm are spun up
even if estimators do not have the tag `tests:core`.

This PR makes behaviour of the `tests:core` tag consistent across
estimators that require VM and those that do not require VM:

a test for a non-core estimator requiring a VM - on its separate VM - is
triggered only if the estimator itself changes.