[ENH] move test skip config to estimators (#8496)
Currently, the information on which tests to skip for a given estimator
is located in `sktime.tests._config`.

This is a design that has a multiple and non-straightforward locus of
extensibility, e.g., one needs to know of the `sktime.tests._config`
file which is not documented.

This PR improves the design by moving the estimator specific config to
the tags of the estimator, to the "CI and tests config" block. This way,
all meta-information about the estimator - including meta-information
related to the testes - is stored in its tags.

It is also more obvious to developers - and users - which tags are
skipped.

This PR changes only the architecture, while leaving most of the skips
still in the `_config` file. Some of the estimators have their
information moved, for testing.