[ENH] The trend argument for ARDL forecaster can now also be 'ctt'. (#8062)
sktime/forecasting/ardl.py provides an sktime layer above the
statsmodels ARDL model.
I recently submitted a PR to statsmodels to allow the trend argument for
ARDL to be 'ctt' (quadratic trend).
The statsmodels PR was accepted and merged into statsmodels main and is
now released.
This sktime PR modifies the documentation of sktime's ARDL to show that
trend can also be 'ctt'.

For reference, here is the statsmodels PR.
https://github.com/statsmodels/statsmodels/pull/9518