[ENH] `tests:libs` tag to specify library dependencies of estimators for differential testing (#8513)
Currently, if an estimator has a business logic dependency in another
module, say, `sktime.my_estimator._other_module`, a change to
`_other_module` does not trigger a test to the estimator.

This is in particular problematic for estimators with dependencies in
the `sktime.libs` folder which contains, for instance, the foundation
model forks and other libraries - since a change in those libraries can
break the estimators depending on them, which are not tested.

This PR adds a `tests:libs` tag which allows to specify the locations of
such important dependencies, triggering the tests.
The tag is added to all reverse dependencies of the `libs` module, in
particular, all foundation models.

The changes in the models should now trigger all tests for
aforementioned estimators, although not directly through the tag.