[ENH] loosen condition for `skpro` dependency exception in `BaseForecaster.predict_proba` (#8554)
Currently, `BaseForecaster.predict_proba` will raise an exception if
`skpro` is not present, and the `_predict_proba` is not the default.

The intention of this is to give users feedback about requiring `skpro`
for non-default implementations of `predict_proba`.

However, this condition is too strict, as some `predict_proba`
implementations, including the default, do not require `skpro`.

Important cases which raise exceptions but would not need to are
composites where `predict_proba` is simply looped through.

This PR ensures that an exception is raised only if the `_predict_proba`
call causes an exception, under the current exception condition, and if
we additionally (new) encounter an `ImportError`.