[BUG] decouple classifier, clusterer, regressor extension contract from `n_jobs` parameter (#8549)
The extension contract for classifiers, clusterers, and regressors
implicitly assumed a `n_jobs` parameter, and that this was written to a
private parameter `self._threads_to_use`.

This assumption is removed, and the onus is now on individual estimators
to handle `n_jobs` and `self._threads_to_use`.

Fixes https://github.com/sktime/sktime/issues/8547, which was added as a regression test.