[BUG] ensure `deps` dependency inspection is compatible with disjunctions in estimator dependency tags (#8585)
This PR fixes an unreported bug, namely that the `deps` dependency
inspection utility would fail if an estimator in the spec contained a
disjunction dependency, for instance, `DartsXGBRegressor`.

This PR fixes the bug by ensuring that disjunctive statements are
resolved by picking the first element, to result in one of multiple
possible choices for dependencies.

One could add a feature to return a disjunctive dependency, but doing
this by default would violate the implicit contract that the return of
`deps` results in a list of PEP 440 dependency specifier str (not a
disjunction, possibly).