[ENH] change `MeanAbsolutePercentageError` parent to `BaseForecastingErrorMetric` (#8632)
Changes the `MeanAbsolutePercentageError` parent to
`BaseForecastingErrorMetric`, from `BaseForecastingErrorMetricFunc`.

Since MAPE is a mean loss, the default `_evaluate` will do the same as
the function imported.

As part of this:

* refactors by moving common internal functions `_percentage_error` and
`_relative_error` to a `_common` module.
* exposes `relative_to` and `eps` as parameters in the metric

This is an architectural improvement since it makes the class self
contained, and avoids dependence on the external function.

Also deduplicates multioutput handling in MAPE and MSE.