[BUG] Fix the predictions from the `ARCH` model implementation (#8610)
#### Reference Issues/PRs

Fixes #8551

#### What does this implement/fix? Explain your changes.

1. Changes focusing in the `_get_arch_result_object()` method :
* Fixed the `start` parameter in the `to_absolute_int()`, where rather
than taking the beginning of `_y`, takes the end of the series
(impacting wrong handling of `fh`)
* Fixed passing value to the `ARCH` horizon parameter in order to
properly retrieve all forecasts to the max horizon provided in `fh`.
    * Fixed the proper indexing of the forecasts.
* The two above bullets ensure also that running forecasts starting with
`fh`>1 works properly.

2. Added `test_arch.py`, as the `ARCH` model did not yet have any file
with tests. One test implemented that checks if the conditional variance
predictions retrieved via `sktime` implementation are matching directly
the `arch` package, moreover with `fh` starting from `2` in order to
test the appropriate indexing of the forecasts, which was previously
faulty.