[BUG] RRF sometimes creates the wrong PeriodIndex when the frequency is higher than 1 (#8606)
#### Reference Issues/PRs

Fixes #8601 

#### What does this implement/fix? Explain your changes.
Makes it so that the created frequencies will always be created at the
start of the frequency instead of at the end.

#### Did you add any tests for the change?
Yes. Functionally identical to the "test_recursive_reduction_with_X" in
the test_reduce file, but the data has a PeriodIndex with frequency "2M"
instead of being numbered.