[MNT] granular test runs in CI in single-estimator VM (#8647)
The single-estimator VM currently run a single test with all tests
dispatched to by `check_estimator`, which makes diagnostics of
individual failures difficult.

This PR splits the single tests into multiple, by adding a
`check_estimator.py` file in the build tools which generates and runs
all tests from a pytest console command, separately.

To test this change, a minimal docstring change is made to
`ChronosForecaster` in order to trigger the VM runs.