[ENH] New Dataset Downloaders for TS Datasets (#8505)
We recently copied TSF and TSC datasets from different sources to
sktime's Hugging Face repository to make downloads more reliable and
reduce the payloads by not downloading datasets in redundant formats.
This PR adds the utility for downloading those datasets from HF repo.
All the public interfaces remain the same.

The `DataDownloader` class follows the strategy pattern and it can be
instantiated with both the HF repo and the fallback URLs, it first tries
to download from the HF repo and upon failure, it tries all the original
sources' URLs.