[MNT] remove `requires_cython` tag and separate VM handling in favour of `tests:vm` (#8674)
This PR removes the `requires_cython` tag and related extra CI by
replacing it with the more broad `tests:vm` tag for the two estimators
affected, `MrSQM` and `MrSEQL`.

Use of the `tests:vm` tag is more configurable, as it allows to install
estimator specific dependencies.

As the `requires_cython` tag was used only in `sktime` internal testing,
no deprecation is required.