[BUG] fix parallelization and pickling in TSC `evaluate` (#8741)
This PR fixes an unreported bug: some parallelization backends in the
TSC `evaluate` would fail (at pickling) since a function was defined in
a local scope, `_evaluate_fold_wrapper`.

However, it appears this function is entirely unnecessary, since all it
does is wrapping a function in a global scope. This PR fixes the bug by
removing the unnecessary wrapper, and point to the function in the
global scope.