[BUG] fix bug in `_safe_import` when importing package and `pkg_name` is not `None` (#8763)
Fixes an unreported bug in `_safe_import` when importing package and
`pkg_name` is not `None` - in this case, the import would always result
in a mock, because the `importlib` statement was referring to the
`pkg_name` and not to `obj_name` which would have been the correct
import name.

A test to prevent regression has also been added.