[DOC] Fix Install Command (#8788)
This PR corrects the installation command in the documentation.

Previously: [`pip install -e
.[dev]`](https://www.sktime.net/en/stable/developer_guide/continuous_integration.html#:~:text=pip%20install%20%2De%20.%5Bdev%5D)

This above command doesn't work in shells like `zsh` due to globbing of
square brackets.

Now: `pip install -e ".[dev]"`

Using quotes makes the command work consistently across different
shells. And this is the way it is used in other places in `sktime` docs
as well.