[ENH] replace `ignores-exogeneous-X` tag by `capability:exogenous` (#8353)
This PR replaces the `ignores-exogeneous-X` tag by a new
`capability:exogenous` tag, which is its negation.

The deprecation is scheduled for version 1.0.0.

In order to ensure consistent aliasing and downwards dependency, the
`_TagAliaserMixin` has been extended to allow a mapping of new/old tags
that is not just name change.

This PR should be merged only with a minor release.