[BUG] Fix `n_instances` test and its values in the tags (#8794)
As discussed in the meet before with @fkiraly , the current test and
values for `n_instances` seems incorrect. `n_instances` implies the
number of time series in a dataset. So, for a panel dataset,
`n_instances = len(df)` is not correct because it counts all the
timepoints across all the instances, and even for dataset which has a
single series, currently `n_instances` is set to number of timepoints.

This PR corrects the test and fixes the values of the tags.