[BUG] fix `PykanForecaster` test failures (#8764)
#### Reference Issues/PRs

Fixes #8735


#### What does this implement/fix? Explain your changes.

The `input_layer_size` was being modified during the fit method but the
updated value wasn't being saved correctly. As a result, the predict
function was using an outdated `input_layer_size`, which was causing the
error.
Updating the `_fit` method to save the change as given below fixes the
issue.
`self._input_layer_size = self.input_layer_size
+X_train.shape[1]*output_size`.