[ENH] Dataset Loader Class for UCR UEA Datasets (#8786)
This PR adds the class for existing dataset loader of UCR UEA datasets
to make them inspectable and easily accessible. I have refrained from
directly inheriting from `_DatasetFromLoaderMixin` class because it
accepts a callable loader function, and does not work for loader
functions like `load_UCR_UEA_dataset` where a string `name` is passed to
them to load the `name` dataset. The logic for the newly added classes
is modified from `_DatasetFromLoaderMixin`.

P.S. This was a part of #8751 but I am splitting them to make the scope
clearer and for my own peace of mind while debugging issues.