[ENH] allow transformer instance as deseasonalizer in `ThetaForecaster`, fix test cases (#8817)
This PR adds the capability to pass a transformer instance to
`ThetaForecaster` as deseasonalizer.

It also fixes the test cases which use multiplicative trend, which
triggers an exception on the test data scenarios on newer `statsmodels`
versions, due to negative values.

As side effect, fixes some current failures on `main`.