[ENH] fixes to tag deprecation logic (#8845)
For 0.39.0, makes fixes in tag deprecation logic, to behave as follows.
Assume a setting where `"old_tag"` is renamed to `"new_tag"`.

If `"new_tag"` is queried and `"old_tag"` is set, retrieves `"old_tag"`.
This is to not break third party estimators that set `"old_tag"`.

If `"new_tag"` is queried and `"old_tag"` is not set, retrieves
`"new_tag"`.

If `"old_tag"` is queried and `"old_tag"` is not set, and `"new_tag"` is
set, retrieves `"new_tag"`. Also raises a warning, since this is a case
where the user queries the tag.

If `"old_tag"` is queried and `"old_tag"` is set, retrieves `"old_tag"`,
and also raises a warning.

Previously, `"new_tag"` would be retrieved it `"new_tag"` was present -
this would break estimators that still had only `"old_tag"` set, but not
`"new_tag"`. With the above, third party extenders have a warning until
1.0.0 to move to the new tags.