[BUG] fixed probabilistic predictions of `SkforecastRecursive` (#8921)
#8255 attempted to resolve #8206 , it did not actually solve. As
@brunoml5 commented at
https://github.com/sktime/sktime/pull/8255#issuecomment-3001524529 and
https://github.com/sktime/sktime/issues/8902#issuecomment-3368514663,
the issue was it was not using the newly used argument and passing to
underlying estimator from `skforecast`. This PR addresses that.

That PR also modified in-sample probabilistic prediction tag and not
out-of-sample tag, which was incorrect. This PR fixes that and closes
#8903 .

Furthermore, this PR reduces number of bins as that was causing dataset
specific errors. It's still a bug on skforecast but it closes #8902 .