[ENH] transformer tags for changes to `unequal_length` property (#8959)
This PR completes the set of tags for transformers describing the change
in the `unequal_length` property of panel data:

* `capability:unequal_length:removes` says that unequal length panels
are all turned into equal length. This tag already existed and was added
with documentation to the tag API reference.
* `capability:unequal_length:adds` says that the transform may turn an
equal length panel to unequal length. This is a new tag needed by
https://github.com/sktime/sktime/pull/8838 and was added.

The PR also adds functionality to the `BaseTransformer` that ensures
that transformations with the `capability:unequal_length:adds` always
return their output in a format supporting unequal length panels.