[BUG] Correct `task_id` in benchmark result for all kinds of dataset loaders (#8897)
_This PR stacks on #8861, please merge it in the order_

Currently the `task_id` or `validation_id` remains blank or more
specifically uses an underscore: `_` in the benchmarking result for
datasets if the loader is a single dataset class or a bundle dataset
class because the name extraction logic is only valid for the loader
functions. This PR fixes it and extracts the name of the dataset from
dataset classes and instances too.