[ENH] Use tirex-ts package for TiRex model (#8876)
#### Reference Issues/PRs
Improves https://github.com/sktime/sktime/pull/8683

#### What does this implement/fix? Explain your changes.
We (NXAI) finally released TiRex as a package to pypi
(https://pypi.org/project/tirex-ts/), what greatly simplifies
integrations like this.

There are also a lot of improvements in the torch/CPU version of TiRex
that make it also easier to run it without complicated environment
setups, that are also integrated by using the newest version of the
package.

#### Does your contribution introduce a new dependency? If yes, which
one?

No, only tirex-ts as an soft dependency of the TiRexForecaster

#### Any other comments?

Always feel free to contact me for any issues related to TiRex
(preferably as an issue on the TiRex repo), as I'm working full time at
NXAI, the company that released TiRex.