[ENH] `is_module_changed` to return `True` if path cannot be found instead of exception raised (#8991)
The `is_module_changed` currently crashes if the path passed to it does
not exist.

This results in test collection failure currently if one of the
`"tests:libs"` paths has a typo or does otherwise not exist.

This PR changes `is_module`changed` to return `True` if path cannot be
found instead of exception raised, ensuring tests collects also for
badly specified `"tests:libs"`.

Not a bug but makes the test system more robust.