[DOC] better documentation for benchmarking file backend handlings json, parquet, csv. (#9043)
Improves and adds documentation for using the benchmarking module and
the `run` method with files.

Adds the backend storage handlers and a description of the file format
to the API documentation for the benchmarking module, for json, parquet,
and csv.

Also makes the null handler (no file written) the default of `run`.