[MNT] Add CITATION.cff for sktime (#9051)
GitHub allows repositories to have a `Cite this repository` button and
the specifications can be included in a `CITATION.cff` file. Please
check out
[https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files](https://docs.github.com/en/repositories/managing-your-repositorys-settings-and-features/customizing-your-repository/about-citation-files)

This PR adds `CITATION.cff` to sktime and adds two identifiers, one for
the original sktime paper and the other one for the Zenodo indexed
version.