[BUG] Improve default hidden layer activation in CNN based classifier and regressor (#9050)
Fixes #9049  

#### What does this implement/fix? Explain your changes.

Changed default hidden layer activation from `sigmoid` to `relu` in
`CNNNetwork`, `CNNClassifier` & `CNNRegressor` class.

#### Any other comments?

In the hidden layers of a CNN, it is non-standard to use sigmoid as an
activation. Sigmoid outputs are in (0,1) and not zero-centred, which can
slow convergence. Also, in general sigmoid is used a final layer
activation in binary classification.

ReLU or a variation of ReLU is a generally agreeable choice for
activation in hidden layers of a CNN.